CFY-2708: Add support for use_external_resource property on the cloudify.vcloud.nodes.Server type.

+++ server_plugin/server.py
@@ -46,6 +46,10 @@ def creation_validation(vca_client, **kwargs):
             if template.get_name() == template_name:
                 return template
 
+    if ctx.node.properties.get('use_external_resource'):
+        # TODO: check: vApp must exists
+        return
+
     server_dict = ctx.node.properties['server']
     required_params = ('catalog', 'template')
     missed_params = set(required_params) - set(server_dict.keys())
@@ -75,6 +79,15 @@ def create(vca_client, **kwargs):
     server.update(ctx.node.properties['server'])
     transform_resource_name(server, ctx)
 
+    if ctx.node.properties.get('use_external_resource'):
+        res_id = ctx.node.properties['resource_id']
+        ctx.instance.runtime_properties[VCLOUD_VAPP_NAME] = res_id
+        ctx.logger.info(
+            "External resource {0} has been used".format(res_id))
+    else:
+        _create(vca_client, config, server)
+    
+def _create(vca_client, config, server):
     vapp_name = server['name']
     vapp_template = server['template']
     vapp_catalog = server['catalog']
@@ -174,16 +187,20 @@ def create(vca_client, **kwargs):
 @operation
 @with_vca_client
 def start(vca_client, **kwargs):
-    vapp_name = get_vapp_name(ctx.instance.runtime_properties)
-    config = get_vcloud_config()
-    vdc = vca_client.get_vdc(config['vdc'])
-    vapp = vca_client.get_vapp(vdc, vapp_name)
-    if _vapp_is_on(vapp) is False:
-        ctx.logger.info("Power-on VApp {0}".format(vapp_name))
-        task = vapp.poweron()
-        if not task:
-            raise cfy_exc.NonRecoverableError("Could not power-on vApp")
-        wait_for_task(vca_client, task)
+    if ctx.node.properties.get('use_external_resource'):
+        ctx.logger.info('not starting server since an external server is '
+                        'being used')
+    else:
+        vapp_name = get_vapp_name(ctx.instance.runtime_properties)
+        config = get_vcloud_config()
+        vdc = vca_client.get_vdc(config['vdc'])
+        vapp = vca_client.get_vapp(vdc, vapp_name)
+        if _vapp_is_on(vapp) is False:
+            ctx.logger.info("Power-on VApp {0}".format(vapp_name))
+            task = vapp.poweron()
+            if not task:
+                raise cfy_exc.NonRecoverableError("Could not power-on vApp")
+            wait_for_task(vca_client, task)
 
     if not _get_state(vca_client):
         return ctx.operation.retry(
@@ -194,29 +211,38 @@ def start(vca_client, **kwargs):
 @operation
 @with_vca_client
 def stop(vca_client, **kwargs):
-    vapp_name = get_vapp_name(ctx.instance.runtime_properties)
-    config = get_vcloud_config()
-    vdc = vca_client.get_vdc(config['vdc'])
-    vapp = vca_client.get_vapp(vdc, vapp_name)
-    ctx.logger.info("Power-off and undeploy VApp {0}".format(vapp_name))
-    task = vapp.undeploy()
-    if not task:
-        raise cfy_exc.NonRecoverableError("Could not undeploy vApp")
-    wait_for_task(vca_client, task)
+    if ctx.node.properties.get('use_external_resource'):
+        ctx.logger.info('not stopping server since an external server is '
+                        'being used')
+    else:
+        vapp_name = get_vapp_name(ctx.instance.runtime_properties)
+        config = get_vcloud_config()
+        vdc = vca_client.get_vdc(config['vdc'])
+        vapp = vca_client.get_vapp(vdc, vapp_name)
+        ctx.logger.info("Power-off and undeploy VApp {0}".format(vapp_name))
+        task = vapp.undeploy()
+        if not task:
+            raise cfy_exc.NonRecoverableError("Could not undeploy vApp")
+        wait_for_task(vca_client, task)
 
 
 @operation
 @with_vca_client
 def delete(vca_client, **kwargs):
-    vapp_name = get_vapp_name(ctx.instance.runtime_properties)
-    config = get_vcloud_config()
-    vdc = vca_client.get_vdc(config['vdc'])
-    vapp = vca_client.get_vapp(vdc, vapp_name)
-    ctx.logger.info("Deleting VApp {0}".format(vapp_name))
-    task = vapp.delete()
-    if not task:
-        raise cfy_exc.NonRecoverableError("Could not delete vApp")
-    wait_for_task(vca_client, task)
+    if ctx.node.properties.get('use_external_resource'):
+        ctx.logger.info('not deleting server since an external server is '
+                        'being used')
+    else:
+        vapp_name = get_vapp_name(ctx.instance.runtime_properties)
+        config = get_vcloud_config()
+        vdc = vca_client.get_vdc(config['vdc'])
+        vapp = vca_client.get_vapp(vdc, vapp_name)
+        ctx.logger.info("Deleting VApp {0}".format(vapp_name))
+        task = vapp.delete()
+        if not task:
+            raise cfy_exc.NonRecoverableError("Could not delete vApp")
+        wait_for_task(vca_client, task)
+
     del ctx.instance.runtime_properties[VCLOUD_VAPP_NAME]
